try fix nuke
